Extracted classes from WindowManager
